Stop populating (and checking) version!
We think version used to exist so we could retrieve events in the order in which
they were received. However, checking and updating it is woefully expensive in
the real world, so we're just not going to anymore.

We can use the database's id field as a proxy for version, since it is
monotonically increasing for every insert; we don't care that there are holes in
the sequence, since all we care is that the elements of the sequence are in order
of insertion.

With postgres (and possibly vertica?) we found that it could take > 30s to check
the last inserted version number before we do a new insert; this is unnecessary
work, since ordering by :id does exactly the same thing as ordering by version.

Solution: stop ordering by version, order by id, and stop populating version.

We should probably remove the version column entirely, but that work is left for
future-us, as soon as we've verified this works and fixes the insert problem.